fix(search/physics/string): mex bounds, impulse, overflow, oob
- RangeMex.Run / MexMaintain.Run / RunWindow: bound mex<64 in the
  advance loops (was: 1L<<64 aliases 1L<<0 -> infinite loop when the
  range contains all of 0..63).
- MexMaintain.RunWindow: drop mex to `remove` when a smaller present
  value leaves the window (was: mex only ever increased -> too large).
- CollisionConstraint.SolveSphere: apply restitution impulse along n
  (jmag = -(1+e)*vn/totalInvMass) before tangential-friction split
  (was: normal velocity written back unchanged -> spheres re-penetrate;
  restitution param accepted but unused).
- IntervalSearch.CountContained: signed int compare (was: uint cast
  wrapped negative starts/points -> wrong counts).
- MedianMaintain.Run: overflow-safe even median lo+(hi-lo)/2 (was:
  (max+min)>>1 overflowed long); return res[n-1] final median (was:
  res[(n-1)/2] median of a prefix).
- SuffixSums.Run<T>/RangeSum<T>: sum float/double as IEEE floats via
  typeof branches (was: reinterpreted bit patterns as int/long ->
  garbage/Inf/NaN).
- TwoPointers.CountPairsWithSum: long sum + count blocks of equal
  duplicates on both arrays (was: int overflow + one pair per match
  undercounted duplicates). HasPairWithSum: long sum (int overflow).
- Lz78.Encode: guard i+pLen>=len (no literal byte) and match only the
  pLen prefix bytes (was: read input[i+pLen] past end when a phrase
  matched to end-of-input).

+++ src/IAFahim.Search.Interval/IntervalSearch.cs
@@ -38,7 +38,7 @@ namespace IAFahim.Search.Interval
             int count = 0;
             for (int i = 0; i < len; i++)
             {
-                if ((uint)starts[i] <= (uint)point && (uint)point < (uint)ends[i])
+                if (starts[i] <= point && point < ends[i])
                 {
                     count++;
                 }
